+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## next
 
 - Changed index and report page buttons to behave like a link (#34)
+- Changed marker of darkmode is enabled from attribute `data-darkmode` to class `.discovery-root-darkmode`
 - Fixed darkmode state restoring from localStorage on init
+- CSS fixes
 
 ## 1.0.0-beta.45 (21-10-2020)